oar/requirements.txt
Django==4.1.2
django-browser-reload
django-tailwind
psycopg2
Markdown==3.4.1
djangorestframework== 3.15.1

